ssn.txt: make it longer and use csv extension
This example is actually too short for a CSV is causing issue with
ChatGPT. ChatGPT throws an error is you send the redacted of the
previous version of the file, even if we are in the middle or not.

This patch just makes it a bit longer so it works fine.